Start working on better algorithms
